Consolidated visualizer code into single script Changed visualizer output path to docs directory Removed old visualize.py from __init__.py
