+++ src/chatbot/__init__.py
@@ -3,4 +3,3 @@ from .repliers import *
 from .types import *
 from .chat import *
 from .cli import *
-from .visualizer import *
